Conflicts:
	Makefile.am
	configure.ac
	include/OSS/Net/IPAddress.h
	include/OSS/SIP/SIPMessage.h
	include/OSS/UTL/PropertyMap.h
	src/Makefile.am
	src/b2bua/SIPB2BScriptableHandler.cpp
	src/b2bua/SIPB2BTransactionManager.cpp
	src/js/JSBase.cpp
	src/unit_test/TestTransport.cpp

+++ src/utl/AdaptiveDelay.cpp
@@ -95,7 +95,7 @@ void AdaptiveDelay::wait()
 
   if (nextWait)
   {
-    timeval sTimeout = { nextWait / 1000000, ( nextWait % 1000000 ) };
+    timeval sTimeout = { (long int)(nextWait / 1000000), (long int)( nextWait % 1000000 ) };
     select( 0, 0, 0, 0, &sTimeout );
   }
 }
